Refactor navigation code (top events)
ViewControllers should be like stupid and not knowing information about the presenter and the presenter should be the one saying "You, ViewController, display this, or stop the spinner, or do whatever!". Presenters should keep the information private in order for the ViewControllers not to access the values of the presenter.